+++ src/HotChocolate/Core/src/Execution/Projections/SelectionExpressionBuilder.cs
@@ -1,12 +1,7 @@
 #if NET8_0_OR_GREATER
-#nullable enable
-
 using System.Collections.Immutable;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq.Expressions;
 using System.Reflection;
-using GreenDonut;
-using GreenDonut.Projections;
 using HotChocolate.Execution.Processing;
 using HotChocolate.Types;
 
@@ -89,26 +84,6 @@ internal sealed class SelectionExpressionBuilder
     {
         public ISelectionSet GetSelectionSet(ISelection selection)
             => Operation.GetSelectionSet(selection, (ObjectType)selection.Type.NamedType());
-    }
-}
-
-[Experimental(Experiments.Projections)]
-public static class HotChocolateExecutionDataLoaderExtensions
-{
-    private static readonly SelectionExpressionBuilder _builder = new();
-
-    public static IDataLoader<TKey, TValue> Select<TKey, TValue>(
-        this IDataLoader<TKey, TValue> dataLoader,
-        ISelection selection)
-        where TKey : notnull
-    {
-        var key = $"{dataLoader.GetType().FullName!}.{selection.Id}";
-        var expression = selection.DeclaringOperation
-            .GetOrAddState<Expression<Func<TValue, TValue>>, SelectionExpressionBuilder>(
-                key,
-                (_, b) => b.BuildExpression<TValue>(selection),
-                _builder);
-        return dataLoader.Select(expression);
     }
 }
 #endif
